monitoring/model_performance/requirements.txt
pandas
alibi-detect
opencensus-ext-azure